ADCs working, mostly
So it turned out we were missing a setting (ADC_CONMODE_bm) that set the
adc to signed mode. Kind of important. Now, adc values are a bit more in
line with what I'd expect. However, they still fluctuate so maybe it
need a filter cap? I also changed the resolution mode to "more than" 12
bit, which lets you use the avgctrl register to play with some built in
averaging and oversampling. It may help, but a simple average function
may do as well and be faster.